requirements.txt
python-dotenv
rich
requests
rocketride
